+++ p2p_wallet/Scenes/Main/BankTransfer/BankTransferCoordinator.swift
@@ -51,41 +51,3 @@ final class BankTransferCoordinator: Coordinator<Void> {
         return controller.deallocatedPublisher().prefix(1).eraseToAnyPublisher()
     }
 }
-
-import SwiftUI
-
-/// In case of successful experiment make a base Renderable protocol
-protocol ChooseItemRenderable<ViewType>: Identifiable where ID == String {
-    associatedtype ViewType: View
-
-    var id: String { get }
-
-    @ViewBuilder func render() -> ViewType
-}
-
-import KeyAppUI
-
-extension Country: ChooseItemRenderable {
-    typealias ViewType = AnyView
-
-    func render() -> AnyView {
-        AnyView(
-            countryView(
-                flag: emoji ?? "",
-                title: name
-            )
-        )
-    }
-
-    private func countryView(flag: String, title: String) -> some View {
-        HStack(spacing: 10) {
-            Text(flag)
-                .font(uiFont: .font(of: .title1, weight: .bold))
-            Text(title)
-                .foregroundColor(Color(Asset.Colors.night.color))
-                .font(uiFont: .font(of: .text3))
-            Spacer()
-        }
-        .padding(.vertical, 6)
-    }
-}
